Added firestore code snippets
