+++ app/src/main/java/com/example/parkease/composables/Cards.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import com.example.parkease.interfaces.ParkingLotData
+import com.example.parkease.utilities.ParkingLotData
 import com.example.parkease.ui.theme.AppTheme
 
 @Composable
