+++ app/src/main/java/com/example/parkease/utilities/FirestoreUtils.kt
@@ -1,0 +1,44 @@
+package com.example.parkease.utilities
+
+import com.google.firebase.firestore.FirebaseFirestore
+
+// Fetch data fields
+// This is a generic function
+fun <T : Any> fetchDocument(
+    collectionName: String,
+    documentId: String,
+    type: Class<T>,
+    onSuccess: (T?) -> Unit,
+    onFailure: (Exception) -> Unit
+) {
+    val db = FirebaseFirestore.getInstance()
+    db.collection(collectionName).document(documentId).get()
+        .addOnSuccessListener { document ->
+            if (document.exists()) {
+                val fields = document.toObject(type)
+                onSuccess(fields)
+            } else {
+                onFailure(Exception("Document does not exist"))
+            }
+        }
+        .addOnFailureListener { exception ->
+            onFailure(exception)
+        }
+}
+
+// Fetch all documents in a collection
+fun fetchCollection(
+    collectionName: String,
+    onSuccess: (List<Map<String, Any>>) -> Unit,
+    onFailure: (Exception) -> Unit
+) {
+    val db = FirebaseFirestore.getInstance()
+    db.collection(collectionName).get()
+        .addOnSuccessListener { result ->
+            val documents = result.map { it.data }
+            onSuccess(documents)
+        }
+        .addOnFailureListener { exception ->
+            onFailure(exception)
+        }
+}
